// _js/snippets/product-options/app/modules/as-enriched-radios/style.module.scss.d.ts
/* DO NOT EDIT: This file is auto-generated and will be overwritten. [build-fingerprint:codegen] */
export type Styles = {
  "ProductRadioButtons": "ProductRadioButtons",
  "ProductRadioButtons-group": "ProductRadioButtons-group",
  "ProductRadioButtons-groupTitle": "ProductRadioButtons-groupTitle",
  "ProductRadioButtons-groupTitle--focused": "ProductRadioButtons-groupTitle--focused",
  "ProductRadioButtons-item": "ProductRadioButtons-item",
  "ProductRadioButtons-item--outOfStock": "ProductRadioButtons-item--outOfStock",
  "ProductRadioButtons-item--selected": "ProductRadioButtons-item--selected",
  "ProductRadioButtons-item--selectedSoldOut": "ProductRadioButtons-item--selectedSoldOut",
  "ProductRadioButtons-item--unavailable": "ProductRadioButtons-item--unavailable",
  "ProductRadioButtons-itemAdditionalInfo": "ProductRadioButtons-itemAdditionalInfo",
  "ProductRadioButtons-itemBadge": "ProductRadioButtons-itemBadge",
  "ProductRadioButtons-itemBenefits": "ProductRadioButtons-itemBenefits",
  "ProductRadioButtons-itemBenefitsIcon": "ProductRadioButtons-itemBenefitsIcon",
  "ProductRadioButtons-itemBenefitsInner": "ProductRadioButtons-itemBenefitsInner",
  "ProductRadioButtons-itemBenefitsItem": "ProductRadioButtons-itemBenefitsItem",
  "ProductRadioButtons-itemBenefitsLabel": "ProductRadioButtons-itemBenefitsLabel",
  "ProductRadioButtons-itemBenefitsList": "ProductRadioButtons-itemBenefitsList",
  "ProductRadioButtons-itemBenefitsWrapper": "ProductRadioButtons-itemBenefitsWrapper",
  "ProductRadioButtons-itemBenefitsWrapper--expanded": "ProductRadioButtons-itemBenefitsWrapper--expanded",
  "ProductRadioButtons-itemBody": "ProductRadioButtons-itemBody",
  "ProductRadioButtons-itemCompareAt": "ProductRadioButtons-itemCompareAt",
  "ProductRadioButtons-itemHeader": "ProductRadioButtons-itemHeader",
  "ProductRadioButtons-itemHeaderBestValue": "ProductRadioButtons-itemHeaderBestValue",
  "ProductRadioButtons-itemInfo": "ProductRadioButtons-itemInfo",
  "ProductRadioButtons-itemInput": "ProductRadioButtons-itemInput",
  "ProductRadioButtons-itemLabel": "ProductRadioButtons-itemLabel",
  "ProductRadioButtons-itemPrice": "ProductRadioButtons-itemPrice",
  "ProductRadioButtons-itemPriceGroup": "ProductRadioButtons-itemPriceGroup",
  "ProductRadioButtons-itemPricing": "ProductRadioButtons-itemPricing",
  "ProductRadioButtons-itemRadioIndicator": "ProductRadioButtons-itemRadioIndicator",
  "ProductRadioButtons-itemServings": "ProductRadioButtons-itemServings",
  "ProductRadioButtons-itemSoldOut": "ProductRadioButtons-itemSoldOut",
  "ProductRadioButtons-itemText": "ProductRadioButtons-itemText",
  "ProductRadioButtons-itemTitle": "ProductRadioButtons-itemTitle",
  "ProductRadioButtons-list": "ProductRadioButtons-list"
}

export type ClassNames = keyof Styles

declare const styles: Styles

export default styles